Fix PAL language poller pointer mask type
The scratch pointer is a Long; the RAM mask must be converted before the bitwise and.

+++ dw2003-dual-screen/app/src/main/java/com/digitaladventure/dw2003/emulation/MemoryPoller.kt
@@ -81,7 +81,7 @@ class MemoryPoller(
         val pointerText = WalkthroughTextFinder.best(
             WalkthroughTextFinder.pointers(scratch).mapNotNull { pointer ->
                 WalkthroughTextFinder.decodeWindow(
-                    read((pointer and RAM_MASK).toInt(), POINTER_WINDOW)
+                    read((pointer and RAM_MASK.toLong()).toInt(), POINTER_WINDOW)
                 )
             }
         )
